+++ packages/core/src/request/index.ts
@@ -1,3 +1,8 @@
 export { MaxTurnsError, PromptCancelledError, ToolApprovalRequiredError } from "./errors";
 export { PromptRequest } from "./prompt-request";
-export type { AgentChildStreamEvent, AgentStreamEvent, PromptResponse } from "./types";
+export type {
+  AgentChildStreamEvent,
+  AgentDeltaEvent,
+  AgentStreamEvent,
+  PromptResponse,
+} from "./types";
